Rewrite

+++ cog_moderation.py
@@ -5,44 +5,65 @@ from discord.ext import commands
 
 
 class Configuration(commands.Cog):
-    """Handles settings like where the bot can post what.
+    """Handles configuration settings like where the bot can post what.
     _client is the bot\'s client
     guild_configs is a dictionary, loaded/saved from guild_configs.json, which keep track of the config settings"""
 
+    default_configuration = { # This is the template for the config objects written to guild_config.json
+                '_guild_alias': '',
+                'bot_channel': -1,
+                'resource_channel': [],
+                'welcome_channel': []
+            }
+
     def __init__(self, _client):
+        """_client is the discord.ext.commands.Bot object which acts as the interface to discord for the bot."""
         self._client = _client
 
         # Load the guild configurations from the json file
         with open(r"resources\guild_configs.json", encoding='utf8') as json_file:
             self.guild_configs = json.load(json_file)
-
-        #TODO upon being added to a guild, add a new config file
-
         return
 
-    def __del__(self):  # Once done with the cog, write the guild configurations to the json file.
-        for guild in self._client.guilds: self.initialize_guild_config(guild) # covering bases.
+    def __del__(self):
+        """When the bot shuts down, write the current configuration object to a json file.
+        Overwrites the previous one to keep track of any changes made while running."""
+
         with open(r"resources\guild_configs.json", 'w', encoding='utf8') as json_file:
-            json.dump(self.guild_configs, json_file)
+            json.dump(self.guild_configs, json_file, indent=1, sort_keys=True)
         return
 
-    def initialize_guild_config(self, guild): # called when a new config entry is needed for a new guild
-        if str(guild.id) not in [*self.guild_configs.keys()]:
-            self.guild_configs[guild.id] = {
-                'bot_channel': -1,
-                'resource_channel': []
-            }
+    def initialize_guild_config(self, guild):
+        """This is called when a guild does not have a config file but needs one.
+        Have to do this weird fruit un-rollup thing [*dict.keys()] because dict.keys() is a naughty boy"""
+
+        if str(guild.id) not in [*self.guild_configs.keys()]: # make sure the guild doesn't already have a config
+            self.guild_configs[guild.id] = self.default_configuration
+            self.guild_configs[guild.id]['_guild_alias'] = guild.name
         return
 
-    async def bot_log(self, context, message):
+    def update_guild_config(self):
+        """Debug method; use this to update the format of the configuration files."""
+        for guild in self._client.guilds:
+            self.guild_configs[str(guild.id)]['_guild_alias'] = guild.name
+        return
+
+    async def bot_log(self, context, message, delete_context_message=True):
+        """Use this to keep stuff clean. Deletes the calling message, prints the desired text."""
+
         bot_channel = context.guild.get_channel(int(self.guild_configs[str(context.guild.id)]['bot_channel']))
         if bot_channel != -1:
-            if isinstance(message, str):
+            if isinstance(message, str): # If the message is a string, we can just print it
                 await bot_channel.send(message)
-            elif isinstance(message, discord.Embed):
+            elif isinstance(message, discord.Embed): # if the message is an embed, use it (use this for images)
                 await bot_channel.send(embed=message)
-        await context.message.delete()
+        if context is not None and delete_context_message: await context.message.delete() # delete the message if wanted
         return
+
+    @commands.Cog.listener()
+    async def on_guild_join(self, guild):
+        """Whenever a new guild is joined, create a configuration for it."""
+        self.initialize_guild_config(guild)
 
     @commands.group(
         name='config',
@@ -50,7 +71,7 @@ class Configuration(commands.Cog):
         help='Bot configuration tools.',
         invoke_without_command=True
     )
-    @commands.has_guild_permissions(manage_guild=True)
+    @commands.has_guild_permissions(view_audit_log=True)
     async def config(self, command_context):
         help_embed = discord.Embed(
             title='[+c | +config] : Configuration Commands',
@@ -75,6 +96,10 @@ class Configuration(commands.Cog):
     )
     async def set(self, command_context, *args):
 
+        # Check to see if the guild is in the list, initialize the config if not.
+        if str(command_context.guild.id) not in [*self.guild_configs.keys()]:
+            self.initialize_guild_config(command_context.guild)
+
         # If no parameters are provided, show what parameters are available to set
         if len(args) == 0:
             help_embed = discord.Embed(
@@ -93,21 +118,40 @@ class Configuration(commands.Cog):
             await self.bot_log(command_context, help_embed)
             return
 
+        #Check to see if the guild is in the list, initialize the config if not.
+        if str(command_context.guild.id) not in [*self.guild_configs.keys()]:
+            self.initialize_guild_config(command_context.guild)
+
+        if args[0] in ['_guild_alias']:
+            await self.bot_log(command_context, f'Error: {args[0]} is not mutable.', False)
+            return
+
         # Check to see if the configuration is allowable.
         if args[0] not in [*self.guild_configs[str(command_context.guild.id)].keys()]:
             await self.bot_log(command_context, f"Statement: ```{args[0]}``` is not a valid configuration parameter.")
+
+        # If marking the bot channel
         elif args[0] == 'bot_channel':
+
+            # If the channel id is the same as the current bot channel, then un-set the bot channel to none.
             if self.guild_configs[str(command_context.guild.id)][args[0]] == command_context.channel.id:
                 self.guild_configs[str(command_context.guild.id)][args[0]] = -1
                 await self.bot_log(command_context, f'{command_context.channel.name} rescinded as the bot channel.')
+
+            # Otherwise overwrite the previous bot channel with the current one
             else:
                 self.guild_configs[str(command_context.guild.id)][args[0]] = command_context.channel.id
                 await self.bot_log(command_context, f'{command_context.channel.name} set as the bot channel.')
 
+        # If marking resource channels
         elif args[0] == 'resource_channel':
+
+            # If the current channel is already marked as such, un-mark it
             if command_context.channel.id in self.guild_configs[str(command_context.guild.id)][args[0]]:
-                del self.guild_configs[str(command_context.guild.id)][args[0]][command_context.channel.id]
+                self.guild_configs[str(command_context.guild.id)][args[0]].remove(command_context.channel.id)
                 await self.bot_log(command_context, f'{command_context.channel.name} removed from resource_channels')
+
+            # Otherwise add it to the list of resource channels.
             else:
                 self.guild_configs[str(command_context.guild.id)][args[0]].append(command_context.channel.id)
                 await self.bot_log(command_context, f'{command_context.channel.name} added to resource_channels')
@@ -130,33 +174,50 @@ class Configuration(commands.Cog):
                     help_embed.add_field(
                         name=key,
                         value='\n'.join([
-                            command_context.guild.get_channel(channel_id).name for channel_id in self.guild_configs[str(command_context.guild.id)][key]
+                            command_context.guild.get_channel(channel_id).name for
+                            channel_id in
+                            self.guild_configs[str(command_context.guild.id)][key]
                         ])
                     )
                 else:
                     help_embed.add_field(
                         name=key,
-                        value= command_context.guild.get_channel(self.guild_configs[str(command_context.guild.id)][key]).name
+                        value=command_context.guild.get_channel(
+                            self.guild_configs[str(command_context.guild.id)][key]
+                        ).name
                     )
 
             await self.bot_log(command_context, help_embed)
         else:
+
+            # If the provided command argument is not one of the available configuration options (+c g arg0 arg1 arg2)
             if args[0] not in [*self.guild_configs[str(command_context.guild.id)].keys()]:
-                await self.bot_log(command_context, f"Statement: ```{args[0]}``` is an invalid configuration parameter.")
+                await self.bot_log(
+                    command_context,
+                    f"Statement: ```{args[0]}``` is an invalid configuration parameter."
+                )
+
+            # if args[0] is in the available configs, then we can display it based on what kind of config it is
             else:
-                description = ''
+
+                # If the config option is a list, list all the things in it
                 if isinstance(self.guild_configs[str(command_context.guild.id)][args[0]], list):
                     description = '\n'.join([
-                        command_context.guild.get_channel(channel_id).name for channel_id in self.guild_configs[str(command_context.guild.id)][args[0]]
+                        command_context.guild.get_channel(channel_id).name for
+                        channel_id in
+                        self.guild_configs[str(command_context.guild.id)][args[0]]
                     ])
+
+                #If the config option is not a list, then just list the thing
                 else:
-                    description = command_context.guild.get_channel(self.guild_configs[str(command_context.guild.id)][args[0]]).name
+                    description = command_context.guild.get_channel(
+                        self.guild_configs[str(command_context.guild.id)][args[0]]
+                    ).name
+
+                # Once the description is made, can send the now-configured embed.
                 await self.bot_log(command_context, discord.Embed(
                     title=args[0],
                     description=description,
                     color=discord.colour.Color.blue()
                 ))
-
         return
-
-# channels: resource, bot, welcome,
